requirements.txt
discord.py
pyarmor
discord.py-Console
colorama